CITRv8: Using Effects and fetching API.
Used useEffect to run a certain request on the
API where it loads the data from the API and
re-render it according to the given array of
dependencies as a second parameter.
If the second parameter is an empty array, then
it will run the request and render it only once.